==> Part 1.ipynb <==
# Project 3 Part 1


## Import

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

akas_url = "data/title.akas.tsv.gz"
akas= pd.read_csv(akas_url, sep='\t', low_memory= False)

basics_url = "data/title.basics.tsv.gz"
basics = pd.read_csv(basics_url, sep='\t', low_memory=False)

ratings_url = "data/title.ratings.tsv.gz"
ratings = pd.read_csv(ratings_url, sep = "\t", low_memory = False)

## Cleaning & Chunking

akas.isna().sum()

basics.isna().sum()

ratings.isna().sum()

akas.replace({'\\N':np.nan},inplace=True)
basics.replace({'\\N':np.nan},inplace=True)
ratings.replace({'\\N':np.nan},inplace=True)

akas

akas = akas[(akas['region']==('US'))].fillna(False)

basics = basics.dropna(subset=['runtimeMinutes','genres','startYear'])

basics['startYear'] = basics['startYear'].astype(float)
## keep startYear 2000-2022
basics = basics[(basics['startYear']>=2000)&(basics['startYear']<2022)]
basics

is_documentary = basics['genres'].str.contains('documentary',case=False)
basics = basics[~is_documentary]

basics = basics[(basics['titleType'] == "movie")]

temp = basics['tconst'].isin(akas['titleId'])
basics = basics[temp]

temp2 = ratings['tconst'].isin(akas['titleId'])
ratings = ratings[temp2]
ratings

# example making new folder with os
import os
os.makedirs('Data/',exist_ok=True) 
# Confirm folder created
os.listdir("Data/")


## Save current dataframe to file.
basics.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)
akas.to_csv("Data/title_akas.csv.gz",compression='gzip',index=False)
ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)

# Open saved file and preview again
basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = True)
basics.info()
basics.head()

akas = pd.read_csv("Data/title_akas.csv.gz", low_memory = True)
akas.info()
akas.head()

ratings = pd.read_csv("Data/title_ratings.csv.gz", low_memory = True)
ratings.info()
ratings.head()

# Cite

 - This data is from https://www.imdb.com/interfaces/ (IMBD)
