==> Part 3.ipynb <==
import numpy as np
import pandas as pd

basics = pd.read_csv('data/title_basics.csv.gz')

## create a col with a list of genres
basics["genres_split"] = basics["genres"].str.split(',')
basics

#Exploding the series and taking the unique values
exploded_genres = basics.explode('genres_split')
exploded_genres

#Taking unique genres only
unique_genres = sorted(exploded_genres['genres_split'].unique())
unique_genres

#saving tconst and genres_split as new df
title_genres = exploded_genres[["tconst","genres_split"]].copy()
title_genres.head()

## Making the genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

## make new integer genre_id and drop string genres
title_genres["genre_id"] = title_genres["genres_split"].map(genre_map)
title_genres = title_genres.drop(columns = "genres_split")

title_genres.head()

# SQLAlchemy

import pymysql
from sqlalchemy import create_engine
from sqlalchemy.types import *
from sqlalchemy_utils import create_database, database_exists
pymysql.install_as_MySQLdb()
username = "root"
password = "root"
db_name = "movies"
connection = f"mysql+pymysql://{username}:{password}@localhost/{db_name}"
    
    
if database_exists(connection):
    print("already exist")
    
else: create_database(connection)

engine = create_engine(connection)

# Basics

basics.info()

basics = basics.drop(columns = ["endYear", "titleType", "originalTitle", "isAdult", "genres", "genres_split"])

## Calculate max string lengths for object columns
key_len = basics['tconst'].fillna('').map(len).max()
title_len = basics['primaryTitle'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_basics = {
    "tconst": String(key_len+1), 
    "primaryTitle": Text(title_len+1),
    'startYear':Float(),
    'runtimeMinutes':Integer()}



# Save to sql with dtype and index=False
basics.to_sql('title_basics',engine,dtype=df_basics, if_exists='replace',index=False)

engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')

f = """SELECT b.tconst, b.primaryTitle, b.startYear, b.runtimeMinutes
FROM title_basics as b
LIMIT 5"""
pd.read_sql(f,engine)

# Title Ratings

title_ratings = pd.read_csv("Data/title_ratings.csv.gz")

title_ratings.info()

## Calculate max string lengths for object columns
key_len = title_ratings['tconst'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_title_ratings = {
    "tconst": String(key_len+1), 
    "averageRating":Float(),
    "numVotes":Integer()}


# Save to sql with dtype and index=False
title_ratings.to_sql('title_ratings',engine,dtype=df_title_ratings, if_exists='replace',index=False)

engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')


g = """SELECT r.tconst, r.averageRating, r.numVotes
FROM title_ratings as r
LIMIT 5"""
pd.read_sql(g,engine)

# The TMDB API Results

tmdb_data = pd.read_csv("Data/tmdb_results_combined.csv.gz")
tmdb_data.info()

tmdb_data = tmdb_data.drop(columns = ["adult", "backdrop_path", "belongs_to_collection", "genres", "homepage", "id",
                                     "original_language", "original_title", "overview", "popularity", "poster_path", 
                                     "production_companies", "production_countries", "release_date", "runtime", 
                                     "spoken_languages", "status", "tagline", "title", "video", "vote_average", 
                                     "vote_count"])

tmdb_data["imdb_id"].duplicated().value_counts()

tmdb_data["imdb_id"] = tmdb_data["imdb_id"].drop_duplicates()
tmdb_data["imdb_id"].duplicated().value_counts()

tmdb_data["imdb_id"].isna().value_counts()

tmdb_data["imdb_id"] = tmdb_data["imdb_id"].fillna('')
tmdb_data["imdb_id"].isna().value_counts()

## Calculate max string lengths for object columns
key_len = tmdb_data['imdb_id'].fillna('').map(len).max()
title_len = tmdb_data['certification'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_tmdb = {
    "imdb_id": String(key_len+1), 
    "revenue":Float(),
    "budget":Float(),
    "certification":Text(title_len+1)}

# Save to sql with dtype and index=False
tmdb_data.to_sql('tmdb_data',engine,dtype=df_tmdb, if_exists='replace',index=False)

engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

h = """SELECT t.imdb_id, t.revenue, t.budget, t.certification 
FROM tmdb_data as t
LIMIT 5"""
pd.read_sql(h,engine)

# Genres

genres = pd.DataFrame({'genre_name': genre_map.keys(),
                       'genre_id': genre_map.values()})

title_len = genres['genre_name'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_genre = {'genre_id': Integer(),
    "genre_name":Text(title_len+1)}

## Set the dataframe index and use index=True 
genres.set_index('genre_id').to_sql('genres',engine,index=True, if_exists='replace')

engine.execute('ALTER TABLE genres ADD PRIMARY KEY (`genre_id`);')

#SQLalchemy
i = """SELECT g.genre_id, genre_name
FROM genres as g
LIMIT 5"""
pd.read_sql(i,engine)

# Tables

q = """SHOW tables"""
pd.read_sql(q, engine)
